+++ packages/gamut-labs/src/experimental/GlobalHeader/types.tsx
@@ -19,6 +19,10 @@ export type User = {
   displayName: string;
   isAdmin?: boolean;
   isCustomerSupport?: boolean;
+  isAccountManager?: boolean;
+  isPaused?: boolean;
+  proTrialCheckoutUrl?: string;
+  showProUpgrade?: boolean;
 };
 
 type LoggedInHeader = BaseHeader & {
@@ -30,6 +34,7 @@ type LoggedInHeader = BaseHeader & {
 export type AnonHeaderVariant = 'landing' | 'login' | 'signup';
 
 export type AnonHeader = BaseHeader & {
+  redirectParam?: string;
   type: 'anon';
   variant?: AnonHeaderVariant;
 };