+++ packages/gamut-labs/src/experimental/index.ts
@@ -1,5 +1,6 @@
 export * from './Card';
 export * from './Coachmark';
+export * from './HubCard';
 export * from './Popover';
 export * from './Text';
 export * from './AppHeader';