+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,163 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [30.5.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.5.0...@codecademy/styleguide@30.5.1) (2021-03-10)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [30.5.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.4.3...@codecademy/styleguide@30.5.0) (2021-03-10)
+
+
+### Features
+
+* **GlobalHeader:** Add pro paused and trial completed header types ([3cf3323](https://github.com/Codecademy/client-modules/commit/3cf33236f7f06457095cd1096949d21622734185))
+
+
+
+### [30.4.3](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.4.2...@codecademy/styleguide@30.4.3) (2021-03-10)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [30.4.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.4.1...@codecademy/styleguide@30.4.2) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [30.4.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.4.0...@codecademy/styleguide@30.4.1) (2021-03-09)
+
+
+### Bug Fixes
+
+* Replace focus-trap with focus-lock-on ([d65e3de](https://github.com/Codecademy/client-modules/commit/d65e3de3149a16603039218c81204a914da6c7cd))
+
+
+
+## [30.4.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.3.0...@codecademy/styleguide@30.4.0) (2021-03-09)
+
+
+### Features
+
+* **Icons:** Adds new icons for Form updates ([5a21007](https://github.com/Codecademy/client-modules/commit/5a21007c5dc0a89a824ad6e561195af78e94513b))
+
+
+
+## [30.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.2.1...@codecademy/styleguide@30.3.0) (2021-03-09)
+
+
+### Features
+
+* **SkipToContent:** SkipToContent + SkipToContent target visual identity + emotion [GM-158] [EGG-701]  ([6e35189](https://github.com/Codecademy/client-modules/commit/6e351891cf499968d2ebd846c36cb299788b0ab5))
+
+
+
+### [30.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.2.0...@codecademy/styleguide@30.2.1) (2021-03-09)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [30.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.1.3...@codecademy/styleguide@30.2.0) (2021-03-08)
+
+
+### Features
+
+* **NotificationItemNew:** map notif type to icon if no image url provided ([7198f23](https://github.com/Codecademy/client-modules/commit/7198f23e64b79ab7b68489c7ea1df11d7c62a704))
+
+
+
+### [30.1.3](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.1.2...@codecademy/styleguide@30.1.3) (2021-03-08)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [30.1.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.1.1...@codecademy/styleguide@30.1.2) (2021-03-05)
+
+
+### Bug Fixes
+
+* **global header:** add support for user role "account manager" ([0cc5ac4](https://github.com/Codecademy/client-modules/commit/0cc5ac4440033dcbe3873658c00def16674fc4ea))
+
+
+
+### [30.1.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.1.0...@codecademy/styleguide@30.1.1) (2021-03-05)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [30.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.0.1...@codecademy/styleguide@30.1.0) (2021-03-05)
+
+
+### Features
+
+* **HubCard:** Created a new hub page card component ([794f369](https://github.com/Codecademy/client-modules/commit/794f369431b1c68191b76b771138275c710997f3))
+
+
+
+### [30.0.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@30.0.0...@codecademy/styleguide@30.0.1) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [30.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@29.7.0...@codecademy/styleguide@30.0.0) (2021-03-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* **Alert:** New Visual Identity Alert [GM-119]
+
+### Features
+
+* **Alert:** New Visual Identity Alert [GM-119] ([c170e6f](https://github.com/Codecademy/client-modules/commit/c170e6ff3e78327ccc6f4875940ea19e26dd65b9))
+
+
+
+## [29.7.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@29.6.9...@codecademy/styleguide@29.7.0) (2021-03-04)
+
+
+### Features
+
+* **ThemeProvider:**  Adds GamutThemeProvider component and initial Global styles ([f4fd13c](https://github.com/Codecademy/client-modules/commit/f4fd13ce04762f75517117372c76d189ee0f0b99))
+
+
+
+### [29.6.9](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@29.6.8...@codecademy/styleguide@29.6.9) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [29.6.8](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@29.6.7...@codecademy/styleguide@29.6.8) (2021-03-04)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
 ### [29.6.7](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@29.6.6...@codecademy/styleguide@29.6.7) (2021-03-04)
 
 **Note:** Version bump only for package @codecademy/styleguide